Update Recursion.cpp
The problem is the while loop if the number is 20 it will overshoot the target because 15 + 6 = 21. The way to fix this is to set the target at 15 and then it will stop before the target. The problem that causes is if the use (0, 2) for the function it will stop at 14. So I made something ahead of add to stop it at 20 with any input I also stop the double display.